OCR 결과를 Block 단위로 추출하기

+++ app/src/main/java/com/eastshine/screentranslator/ocr/model/TextElement.kt
@@ -7,7 +7,7 @@ data class TextElement(
     val text: String,
     val boundingBox: Rect,
     val cornerPoints: Array<Point>?,
-    val confidence: Float
+    val confidence: Double,
 ) {
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
